python/requirements.txt
tensorflow
matplotlib
pillow
opencv-python
keras
scipy